pkg/op-forum-agg/pyproject.toml
[tool.poetry]
name = "op-forum-agg"
version = "0.1.0"
description = ""
authors = ["devjoaov <joaovictorass95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-cors = "^4.0.1"
flask-limiter = "^3.8.0"
httpx = "^0.27.0"
psycopg2 = "^2.9.9"
cchardet = "2.2.0a2"
redis = "^5.0.7"
celery = "^5.4.0"
flower = "^2.0.1"
forum-dl = "^0.3.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

